requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.23.5
matplotlib==3.7.1
tqdm==4.65.0
scikit-learn==1.2.2
tensorboard==2.13.0
opencv-python==4.8.0.74
pillow==9.5.0 